backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
crewai==0.28.0
langchain==0.0.335
langchain-community==0.0.13
google-search-results==2.4.2
openai==1.3.5 